recommendation-engine/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.9.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
frozenlist==1.5.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.3.1
marshmallow==3.25.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy>=1.22.4,<2
openai==1.59.5
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xlrd==2.0.1
yarl==1.18.3
zstandard==0.23.0
celery==5.5.0